Test and refactor source
