--- .npmrc
@@ -1,1 +1,0 @@
-@github:registry=http://npm.pkg.github.com/:_authToken=${NPM_AUTH_TOKEN}
