+++ jest.config.d.ts
@@ -1,2 +1,2 @@
-declare const _exports: import("ts-jest/dist/types").InitialOptionsTsJest;
-export = _exports;
+declare const _exports: import('ts-jest/dist/types').InitialOptionsTsJest
+export = _exports
