+++ package.json
@@ -7,7 +7,7 @@
     "build": "tsc",
     "format": "prettier --write '**/*.ts'",
     "format-check": "prettier --check '**/*.ts'",
-    "lint": "eslint --fix src/**/*.ts test/**/*.ts",
+    "lint": "eslint --fix src/**/*.ts",
     "package": "ncc build --source-map --license licenses.txt",
     "test": "jest",
     "test:watch": "jest --watch src",
@@ -25,21 +25,25 @@
   },
   "homepage": "https://github.com/dsp-testing/go-snapshot-action#readme",
   "devDependencies": {
+    "@types/jest": "^27.5.2",
     "@typescript-eslint/eslint-plugin": "^5.20.0",
     "@typescript-eslint/parser": "^5.20.0",
     "@vercel/ncc": "^0.33.4",
     "eslint": "^8.13.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.5.3",
     "eslint-plugin-n": "^15.1.0",
     "eslint-plugin-promise": "^6.0.0",
     "jest": "^28.0.0",
-    "ts-jest": "^28.0.2",
+    "ts-jest": "^28.0.4",
     "typescript": "^4.6.4"
   },
   "dependencies": {
     "@actions/core": "^1.6.0",
+    "@actions/exec": "^1.1.1",
     "@actions/github": "^5.0.0",
-    "@github/dependency-submission-toolkit": "^1.1.1"
+    "@github/dependency-submission-toolkit": "^1.1.2",
+    "packageurl-js": "^0.0.6"
   }
 }
