+++ src/index.ts
@@ -1,27 +1,12 @@
-import * as path from 'path'
-import * as process from 'process'
+import path from 'path'
 import fs from 'fs'
 
-import * as exec from '@actions/exec'
 import * as core from '@actions/core'
-import {
-  Manifest,
-  BuildTarget,
-  PackageCache,
-  Snapshot,
-  submitSnapshot
-} from '@github/dependency-submission-toolkit'
+import { Snapshot, submitSnapshot } from '@github/dependency-submission-toolkit'
 
-import { parseGoList, parseGoModGraph } from './parse-go-package'
+import { processGoGraph, processGoBuildTarget } from './process'
 
-// Set the detector information provided from the action workflow input
-const detector = {
-  name: core.getInput('detector-name'),
-  url: core.getInput('detector-url'),
-  version: core.getInput('detector-version')
-}
-
-async function main() {
+async function main () {
   const goModPath = path.normalize(core.getInput('go-mod-path'))
 
   if (path.basename(goModPath) !== 'go.mod' && fs.existsSync(goModPath)) {
@@ -50,7 +35,11 @@ async function main() {
   }
 
   const packageCache = await processGoGraph(goModDir)
-  const manifest = await processGoTarget(goModDir, goBuildTarget, packageCache)
+  const manifest = await processGoBuildTarget(
+    goModDir,
+    goBuildTarget,
+    packageCache
+  )
   const snapshot = new Snapshot({
     name: 'github-go-dependency-detector',
     url: 'https://github.com/github/github-go-dependency-detector',
@@ -58,62 +47,6 @@ async function main() {
   })
   snapshot.addManifest(manifest)
   submitSnapshot(snapshot)
-}
-
-async function processGoGraph(goModDir: string): Promise<PackageCache> {
-  console.log(`Running 'go mod graph' in ${goModDir}`)
-  const goModGraph = await exec.getExecOutput('go', ['mod', 'graph'], {
-    cwd: goModDir
-  })
-  if (goModGraph.exitCode !== 0) {
-    core.error(goModGraph.stderr)
-    core.setFailed("'go mod graph' failed!")
-    throw new Error("Failed to execute 'go mod graph'")
-  }
-  const cache = new PackageCache()
-  const packageAssocList = parseGoModGraph(goModGraph.stdout)
-  packageAssocList.forEach(([parentPkg, childPkg]) => {
-    cache.package(parentPkg).dependsOn(cache.package(childPkg))
-  })
-  return cache
-}
-
-// For a specific Go _build target_, this template lists all dependencies used
-// to build the build target It does not provide association between the
-// dependencies (i.e. which dependencies depend on which)
-// eslint-disable-next-line quotes
-// eslint-disable-next-line no-useless-escape
-const goListDependencyTemplate =
-  '"{{define \\"M\\"}}{{.Path}}@{{.Version}}{{end}}{{with .Module}}{{if not .Main}}{{if .Replace}}{{template \\"M\\" .Replace}}{{else}}{{template \\"M\\" .}}{{end}}{{end}}{{end}}"'
-async function processGoTarget(
-  goModDir: string,
-  goBuildTarget: string,
-  cache: PackageCache
-): Promise<Manifest> {
-  console.log(
-    `Running go package detection in ${goModDir} on build target ${goBuildTarget}`
-  )
-
-  const goList = await exec.getExecOutput(
-    'go',
-    ['list', '-deps', '-f', goListDependencyTemplate, goBuildTarget],
-    { cwd: goModDir }
-  )
-  if (goList.exitCode !== 0) {
-    core.error(goList.stderr)
-    core.setFailed("'go list' failed!")
-    throw new Error("Failed to execute 'go list'")
-  }
-
-  const dependencies = parseGoList(goList.stdout)
-  const manifest = new BuildTarget(
-    goBuildTarget,
-    path.join(goModDir, goBuildTarget)
-  )
-  dependencies.forEach((dep) => {
-    manifest.addBuildDependency(cache.package(dep))
-  })
-  return manifest
 }
 
 main()
